SQL-like cascade delete
